Sushi/requirements.txt
NMAP needs to be installed.

python library dependancies:

    python-pyfiglet
    python-NMAP
    